Move Autoconsent Pixel def to BSK
